+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/amazon-bedrock
 
+## 4.0.52
+
+### Patch Changes
+
+- Updated dependencies [0a0d29c]
+  - @ai-sdk/anthropic@3.0.39
+
+## 4.0.51
+
+### Patch Changes
+
+- Updated dependencies [7168375]
+  - @ai-sdk/provider@3.0.8
+  - @ai-sdk/anthropic@3.0.38
+  - @ai-sdk/provider-utils@4.0.14
+
 ## 4.0.50
 
 ### Patch Changes